src/data-pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.15.2",
    "internetarchive>=5.7.2",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "sarvamai>=0.1.24",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]
